cloud-functions/send-certificate-emails/pyproject.toml
[project]
name = "send-certificate-emails"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "functions-framework>=3.10.1",
    "google-auth>=2.49.1",
    "google-cloud-storage>=3.9.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "pydantic"
]